% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreement.R
\name{agreement}
\alias{agreement}
\title{Agreement Sensor}
\usage{
agreement(z1, z2, eps = 1e-08)
}
\arguments{
\item{z1}{First representation}

\item{z2}{Second representation}

\item{eps}{Small constant}
}
\description{
Agreement Sensor
}
