% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reslik.R
\name{reslik}
\alias{reslik}
\title{Population-Level Sensor (ResLik)}
\usage{
reslik(z, ref_mean = 0, ref_sd = 1, lambda = 1, tau = 0.05)
}
\arguments{
\item{z}{Numeric vector or matrix.}

\item{ref_mean}{Numeric. Reference mean.}

\item{ref_sd}{Numeric. Reference standard deviation.}

\item{lambda}{Numeric. Sensitivity.}

\item{tau}{Numeric. Dead-zone threshold.}
}
\description{
Population-Level Sensor (ResLik)
}
