% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcs.R
\name{tcs}
\alias{tcs}
\title{Temporal Consistency Sensor}
\usage{
tcs(z_t, z_prev, eps = 1e-06)
}
\arguments{
\item{z_t}{Current representation (vector or matrix)}

\item{z_prev}{Previous representation (same shape)}

\item{eps}{Small constant}
}
\description{
Temporal Consistency Sensor
}
