% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_surface.R
\name{rlcs_control}
\alias{rlcs_control}
\title{Deterministic Control Surface}
\usage{
rlcs_control(reslik, tcs = NULL, agreement = NULL, thresholds = list())
}
\arguments{
\item{reslik}{Output from reslik()}

\item{tcs}{Output from tcs() (optional)}

\item{agreement}{Output from agreement() (optional)}

\item{thresholds}{List of thresholds}
}
\description{
Deterministic Control Surface
}
